Desafio IA - Resolvido/requirements.txt
langchain
langchain-google-genai
chromadb
pypdf
python-dotenv
streamlit
google-generativeai
